Moved modal and delete logic
